requirements.txt
Flask
pandas
scikit-learn
google-cloud-bigquery
mlxtend
gunicorn
db-dtypes
